+++ .gitignore
@@ -19,3 +19,4 @@ package-lock.json
 # IDE Sharing
 .floo
 .flooignore
+.vscode
