+++ .jest/setup.js
@@ -1,0 +1,8 @@
+window.matchMedia = window.matchMedia ||
+  function watchMedia () {
+    return {
+      matches: false,
+      addListener: () => {},
+      removeListener: () => {},
+    }
+  }
