+++ src/components/DropdownContent/index.js
@@ -1,21 +1,3 @@
-import React, { PureComponent } from 'react'
+import DropdownContent from './DropdownContent'
 
-import { Consumer } from '../Dropdown'
-import DropdownContentControlled from '../DropdownContentControlled'
-
-
-export default class DropdownContent extends PureComponent {
-  render () {
-    return (
-      <Consumer>
-        {({ show, toggle }) =>
-          <DropdownContentControlled
-            show={show}
-            onClose={toggle}
-            {...this.props}
-          />
-        }
-      </Consumer>
-    )
-  }
-}
+export default DropdownContent
