+++ src/components/DropdownFooter/index.js
@@ -1,32 +1,3 @@
-import React, { PureComponent } from 'react'
-import PropTypes from 'prop-types'
-import cn from 'classnames'
+import DropdownFooter from './DropdownFooter'
 
-import { PROP_TYPE_CHILDREN } from '../constants'
-
-
-export default class DropdownHeader extends PureComponent {
-  static propTypes = {
-    children: PROP_TYPE_CHILDREN,
-    className: PropTypes.string,
-  }
-
-  render () {
-    const {
-      children,
-      className,
-      ...restProps
-    } = this.props
-
-    const classes = cn({
-      'mc-dropdown__footer': true,
-      [className]: className,
-    })
-
-    return (
-      <div className={classes} {...restProps}>
-        {children}
-      </div>
-    )
-  }
-}
+export default DropdownFooter
