+++ src/components/DropdownToggle/index.js
@@ -1,27 +1,3 @@
-import React, { PureComponent } from 'react'
+import DropdownToggle from './DropdownToggle'
 
-import { Consumer } from '../Dropdown'
-import { PROP_TYPE_CHILDREN } from '../constants'
-import { renderChildren } from '../helpers'
-
-
-export default class DropdownToggle extends PureComponent {
-  static propTypes = {
-    children: PROP_TYPE_CHILDREN,
-  }
-
-  render () {
-    const { children } = this.props
-
-    return (
-      <Consumer>
-        {({ toggle, toggleRef }) =>
-          renderChildren(children, {
-            onClick: toggle,
-            ref: toggleRef,
-          })
-        }
-      </Consumer>
-    )
-  }
-}
+export default DropdownToggle
