+++ src/components/index.js
@@ -9,6 +9,14 @@ export { default as CarouselConnector } from './CarouselConnector'
 export { default as Checkbox } from './Checkbox'
 export { default as CheckboxField } from './CheckboxField'
 export { default as ClickOutside } from './ClickOutside'
+export { default as Dropdown } from './Dropdown'
+export { default as DropdownBody } from './DropdownBody'
+export { default as DropdownContent } from './DropdownContent'
+export { default as DropdownContentControlled } from './DropdownContentControlled'
+export { default as DropdownFooter } from './DropdownFooter'
+export { default as DropdownHeader } from './DropdownHeader'
+export { default as DropdownItem } from './DropdownItem'
+export { default as DropdownToggle } from './DropdownToggle'
 export { default as FormGroup } from './FormGroup'
 export { default as HoverHandler } from './HoverHandler'
 export { default as Icon } from './Icons'
